fixed issues with createEvents
basically fixed the create events added a way to create meeting times and finally add physicality, also added a condition for also register in if it there.

have to write functions for write to different json files and some other supporting function for it.

changed the way I add staff so that I can just id the staff id which I already have in other objects.

removed the get all meeting days now it is handled by the createEvents.

changed the way I return in parseMeetingDay to a type so that it is easier to access.

added a populate course so that I can populate it with the class info.