Changed Addstaff, add dept and some more things
Basically I wanted to be able to have add staff in the appropriate folder which was not devided by the semesters ad this required some change in the code. so fixed that. Now I am finally able to just add the loops and then I should be finally done with this.